requirements.txt
openai
langchain
serpapi
streamlit
requests
python-dotenv
